+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/lib/src/method_channel/method_channel_google_maps_flutter.dart
@@ -437,6 +437,13 @@ class MethodChannelGoogleMapsFlutter extends GoogleMapsFlutterPlatform {
     return channel(mapId).invokeMethod<double>('map#getZoomLevel');
   }
 
+  /// Converts a distance in meters to content size.
+  Future<double> getPoint(double meters, LatLng latLng, {@required int mapId}) {
+    assert(meters != null && latLng != null);
+    return channel(mapId).invokeMethod<double>('map#getPoint',
+        <String, dynamic>{"meter": meters, "location": latLng.toJson()});
+  }
+
   /// Returns the image bytes of the map
   @override
   Future<Uint8List> takeSnapshot({