+++ packages/google_maps_flutter/google_maps_flutter_platform_interface/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 1.0.2
+
+* Support projection method pointForMeters, to get the distance in meters to content size from the specified LatLng in the map for iOS.
+
 ## 1.0.1
 
 * Initial open source release.